requirements.txt
pymupdf==1.23.8
faiss-cpu==1.7.4
python-dotenv==1.0.1
groq==0.26.0
langchain==0.1.12
langchain-groq==0.0.1
langchain-community>=0.0.28,<0.1.0
scikit-learn==1.4.1.post1
streamlit==1.32.2
requests==2.31.0 